requirements.txt
flask
numpy
requests
psycopg2-binary
python-dotenv
asyncua
streamlit
plotly
pandas